把不同的kernal獨立成不同的puy檔
